+++ src/views/BasePage.vue
