+++ src/views/DetailPage.vue
