+++ src/views/NewPaymentView.vue
@@ -3,7 +3,6 @@ import VgInput from '@/components/ui/forms/VgInput.vue'
 import VgInputNumber from '@/components/ui/forms/VgInputNumber.vue'
 import { computed, reactive } from 'vue'
 import VgButton from '@/components/ui/VgButton.vue'
-import VgInputCheck from '@/components/ui/forms/VgInputCheck.vue'
 import VgInputSelect from '@/components/ui/forms/VgInputSelect.vue'
 import VgInputSwitch from '@/components/ui/forms/VgInputSwitch.vue'
 import { useGlobalStore } from '@/stores/global.store'
